In the end, added ForeignVocabQuizGUI back. Isolated part of the code that doesn't work. likely due to action listener. Could not get game to work through the menu. Game only works hard coded at the moment. When combined with menu + readin file working, the game itself can not be played. Will edit readme file to add things found out
