+++ src/menu.java
@@ -14,38 +14,7 @@ public class menu
 	private JFrame frame1, frame2; 
 	private ActionListener action; 
 	private JButton spanish, french, german; 
-	private final static String newLine = "\n";
-   	private JTextField field;
-   	private JLabel label3; //arbitrary name, should fix it
-    	private ForeignVocabQuiz quiz;
-    	private String userGuess;
-    	private String word;
-    	private String counterPart;
-  	private int totalQuestions;
-    	private int questionsCorrect;
-    	private int numOfGuesses;	
-
-	
-	 class Listener implements ActionListener {
-	public void actionPerformed(ActionEvent event) {
-	    userGuess = field.getText();
-	    if(numOfGuesses < 3){
-		if(quiz.checkUserGuess(userGuess))
-		    label3.setText("Correct!");		
-		else{
-		    if (numOfGuesses == 2)
-			label3.setText("The correct answer was: " + counterPart);
-		    else
-			label3.setText("False");
-		    numOfGuesses++;
-		}
-	    }
-	    field.requestFocus();
-	    field.selectAll();
-	}
-    }
-
-	public void createAndDisplayGUI()
+public void createAndDisplayGUI()
 	{
 		frame1 = new JFrame("Menu GUI"); 
 		frame1.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
@@ -69,67 +38,32 @@ public class menu
 			public void actionPerformed(ActionEvent e)
 			{
 				String command = e.getActionCommand(); 
-				quiz = new ForeignVocabQuiz(); 
-				/*
 				if(command.equals("german"))
 				{
-					quiz = new ForeignVocabQuiz(); 
+					ForeignVocabQuizGUI gui =  new ForeignVocabQuizGUI(); 
+					gui.go(); 
 				}
 				if(command.equals("spanish"))
 				{
-					quiz = new ForeignVocabQuiz("spanish.txt");
+					ForeignVocabQuizGUI gui = new ForeignVocabQuizGUI("spanish.txt");
+					gui.go();
 				}
 				if(command.equals("french"))
 				{
-					quiz = new ForeignVocabQuiz("french.txt");
+					ForeignVocabQuizGUI gui = new ForeignVocabQuizGUI("french.txt");
+					gui.go();
 				}
-				*/
-				frame2 = new JFrame("Foreign Vocab Quiz GUI");
-				frame2.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE); 
-				frame2.setLocationByPlatform(true);
-				JTextArea text = new JTextArea(10,35);
-				field = new JTextField(20);
-				JLabel label = new JLabel("Your answer: ");
-				JLabel label2 = new JLabel();
-				label3 = new JLabel();
-				JLabel label4 = new JLabel("Your score is " + questionsCorrect + "/" + totalQuestions + ".");
-				JButton button = new JButton("answer");
-				JPanel panel = new JPanel();
-  				JPanel panel2 = new JPanel();
-				JPanel panel3 = new JPanel();
-				JScrollPane scroller = new JScrollPane(text);
-				scroller.setVerticalScrollBarPolicy(ScrollPaneConstants.VERTICAL_SCROLLBAR_ALWAYS);
-				scroller.setHorizontalScrollBarPolicy(ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
-				text.setLineWrap(true);
-				text.setEditable(false);
-				text.append("Welcome!" + newLine);
-				text.append("This is a vocabulary quiz for a foreign language." + newLine);
-				text.append("You have three chances per word." + newLine);
-				text.append("We'll begin now." + newLine);
-				field.requestFocus();
-				field.addActionListener(new Listener());
-				button.addActionListener(new Listener());
-				panel2.add(label2);
-				panel.add(label);
-				panel.add(field);
-				panel.add(button);
-				panel.add(label3);
-				panel3.add(label4);
-				panel3.add(scroller);
-				frame2.getContentPane().add(BorderLayout.NORTH, panel2);
-				frame2.getContentPane().add(BorderLayout.CENTER, panel);
-				frame2.getContentPane().add(BorderLayout.SOUTH, panel3);
-				frame2.setSize(450,300);
-				frame2.setVisible(true);
-				frame1.setVisible(false); 
-				frame1.dispose();
+                	
+
+             
 			} 
 		};
 	german.addActionListener(action);
 	french.addActionListener(action);
 	spanish.addActionListener(action);
-
-    }					
+	
+}
+				
 	public static void main(String [] args){
 		SwingUtilities.invokeLater(new Runnable()
 		{
